test: Added tests for parse_via_chat_gpt
