+++ test/glific/flows/webhook_test.exs
@@ -10,13 +10,9 @@ defmodule Glific.Flows.WebhookTest do
   }
 
   alias Glific.{
-    Clients.CommonWebhook,
     Fixtures,
-    Messages,
     Seeds.SeedsDev
   }
-
-  import Mock
 
   setup do
     default_provider = SeedsDev.seed_providers()
@@ -261,290 +257,5 @@ defmodule Glific.Flows.WebhookTest do
 
       assert WebhookLog.count_webhook_logs(%{filter: attrs}) == logs_count + 1
     end
-  end
-
-  test "successful geolocation response" do
-    lat = "37.7749"
-    long = "-122.4194"
-    fields = %{"lat" => lat, "long" => long}
-
-    Tesla.Mock.mock(fn
-      %{method: :get} ->
-        %Tesla.Env{
-          status: 200,
-          body:
-            Jason.encode!(%{
-              "results" => [
-                %{
-                  "address_components" => [
-                    %{"long_name" => "San Francisco", "types" => ["locality"]},
-                    %{"long_name" => "CA", "types" => ["administrative_area_level_1"]},
-                    %{"long_name" => "USA", "types" => ["country"]}
-                  ],
-                  "formatted_address" => "San Francisco, CA, USA"
-                }
-              ]
-            })
-        }
-    end)
-
-    result = CommonWebhook.webhook("geolocation", fields)
-
-    assert result[:success] == true
-    assert result[:city] == "San Francisco"
-    assert result[:state] == "CA"
-    assert result[:country] == "USA"
-    assert result[:postal_code] == "N/A"
-    assert result[:district] == "N/A"
-    assert result[:ward] == "N/A"
-    assert result[:address] == "San Francisco, CA, USA"
-  end
-
-  test "geolocation failure response" do
-    lat = "37.7749"
-    long = "-122.4194"
-    fields = %{"lat" => lat, "long" => long}
-
-    # Mock a non-200 response from the API (e.g., 500 Internal Server Error)
-    Tesla.Mock.mock(fn
-      %{method: :get} ->
-        %Tesla.Env{
-          status: 500,
-          body: "Internal Server Error"
-        }
-    end)
-
-    result = CommonWebhook.webhook("geolocation", fields)
-
-    # Assert that success is false and an error message is returned
-    refute result[:success]
-    refute is_nil(result[:error])
-    assert result[:error] == "Received status code 500"
-  end
-
-  @tag :respo
-  test "parse_via_gpt_vision without response_format params, trying to get valid json" do
-    with_mock(
-      Messages,
-      validate_media: fn _, _ -> %{is_valid: true, message: "success"} end
-    ) do
-      Tesla.Mock.mock(fn
-        %{url: "https://api.openai.com/v1/chat/completions"} ->
-          %Tesla.Env{
-            status: 200,
-            body: %{
-              "choices" => [
-                %{
-                  "message" => %{
-                    "content" => "```json\n{\n  \"steps\": 4,\n  \"answer\": 10\n}\n```"
-                  }
-                }
-              ]
-            }
-          }
-      end)
-
-      fields = %{
-        "prompt" =>
-          "ignore the image, value of steps is 4 and value of answer is 10, give in valid json",
-        "url" =>
-          "https://fastly.picsum.photos/id/145/200/300.jpg?hmac=mIsOtHDzbaNzDdNRa6aQCd5CHCVewrkTO5B1D4aHMB8",
-        "model" => "gpt-4o"
-      }
-
-      assert %{success: true, response: "```json\n{\n  \"steps\": 4,\n  \"answer\": 10\n}\n```"} =
-               CommonWebhook.webhook("parse_via_gpt_vision", fields)
-    end
-  end
-
-  @tag :respo
-  test "parse_via_gpt_vision with response_format params type json_object, trying to get valid json" do
-    with_mock(
-      Messages,
-      validate_media: fn _, _ -> %{is_valid: true, message: "success"} end
-    ) do
-      Tesla.Mock.mock(fn
-        %{url: "https://api.openai.com/v1/chat/completions"} ->
-          %Tesla.Env{
-            status: 200,
-            body: %{
-              "choices" => [
-                %{
-                  "message" => %{
-                    "content" => "{\n  \"steps\": 4,\n  \"answer\": 10\n}"
-                  }
-                }
-              ]
-            }
-          }
-      end)
-
-      fields = %{
-        "prompt" =>
-          "ignore the image, value of steps is 4 and value of answer is 10, give in valid json",
-        "url" =>
-          "https://fastly.picsum.photos/id/145/200/300.jpg?hmac=mIsOtHDzbaNzDdNRa6aQCd5CHCVewrkTO5B1D4aHMB8",
-        "model" => "gpt-4o",
-        "response_format" => %{"type" => "json_object"}
-      }
-
-      assert %{success: true, response: %{"steps" => 4, "answer" => 10}} =
-               CommonWebhook.webhook("parse_via_gpt_vision", fields)
-    end
-  end
-
-  @tag :respo
-  test "parse_via_gpt_vision with invalid response_format param, trying to get valid json" do
-    with_mock(
-      Messages,
-      validate_media: fn _, _ -> %{is_valid: true, message: "success"} end
-    ) do
-      Tesla.Mock.mock(fn
-        %{url: "https://api.openai.com/v1/chat/completions"} ->
-          %Tesla.Env{
-            status: 200,
-            body: %{
-              "choices" => [
-                %{
-                  "message" => %{
-                    "content" => "{\n  \"steps\": 4,\n  \"answer\": 10\n}"
-                  }
-                }
-              ]
-            }
-          }
-      end)
-
-      fields = %{
-        "prompt" =>
-          "ignore the image, value of steps is 4 and value of answer is 10, give in valid json",
-        "url" =>
-          "https://fastly.picsum.photos/id/145/200/300.jpg?hmac=mIsOtHDzbaNzDdNRa6aQCd5CHCVewrkTO5B1D4aHMB8",
-        "model" => "gpt-4o",
-        # the response format is invalid
-        "response_format" => %{"type" => "json_objectz"}
-      }
-
-      assert "response_format type should be json_schema or json_object" =
-               CommonWebhook.webhook("parse_via_gpt_vision", fields)
-    end
-  end
-
-  @tag :respo
-  test "parse_via_gpt_vision with response_format param as json_schema, trying to get valid json" do
-    with_mock(
-      Messages,
-      validate_media: fn _, _ -> %{is_valid: true, message: "success"} end
-    ) do
-      Tesla.Mock.mock(fn
-        %{url: "https://api.openai.com/v1/chat/completions"} ->
-          %Tesla.Env{
-            status: 200,
-            body: %{
-              "choices" => [
-                %{
-                  "message" => %{
-                    "content" => "{\n  \"steps\": \"4\",\n  \"answer\": \"10\"\n}"
-                  }
-                }
-              ]
-            }
-          }
-      end)
-
-      fields = %{
-        "prompt" =>
-          "ignore the image, value of steps is 4 and value of answer is 10, give in valid json",
-        "url" =>
-          "https://fastly.picsum.photos/id/145/200/300.jpg?hmac=mIsOtHDzbaNzDdNRa6aQCd5CHCVewrkTO5B1D4aHMB8",
-        "model" => "gpt-4o",
-        "response_format" => %{
-          "type" => "json_schema",
-          "json_schema" => %{
-            "name" => "schemaing",
-            "strict" => true,
-            "schema" => %{
-              "type" => "object",
-              "properties" => %{
-                "steps" => %{
-                  "type" => "string"
-                },
-                "answer" => %{
-                  "type" => "string"
-                }
-              },
-              "required" => [
-                "steps",
-                "answer"
-              ],
-              "additionalProperties" => false
-            }
-          }
-        }
-      }
-
-      assert %{success: true, response: %{"steps" => "4", "answer" => "10"}} =
-               CommonWebhook.webhook("parse_via_gpt_vision", fields)
-    end
-  end
-
-  @tag :respo
-  test "parse_via_gpt_vision with response_format param as invalid json_schema, trying to get valid json" do
-    with_mock(
-      Messages,
-      validate_media: fn _, _ -> %{is_valid: true, message: "success"} end
-    ) do
-      Tesla.Mock.mock(fn
-        %{url: "https://api.openai.com/v1/chat/completions"} ->
-          %Tesla.Env{
-            status: 400,
-            body: %{
-              "error" => %{
-                "message" =>
-                  "Invalid schema for response_format 'schemaing': In context=(), 'additionalProperties' is required to be supplied and to be false."
-              }
-            }
-          }
-      end)
-
-      fields = %{
-        "prompt" =>
-          "ignore the image, value of steps is 4 and value of answer is 10, give in valid json",
-        "url" =>
-          "https://fastly.picsum.photos/id/145/200/300.jpg?hmac=mIsOtHDzbaNzDdNRa6aQCd5CHCVewrkTO5B1D4aHMB8",
-        "model" => "gpt-4o",
-        "response_format" => %{
-          "type" => "json_schema",
-          "json_schema" => %{
-            "name" => "schemaing",
-            "strict" => true,
-            "schema" => %{
-              "type" => "object",
-              "properties" => %{
-                "steps" => %{
-                  "type" => "string"
-                },
-                "answer" => %{
-                  "type" => "string"
-                }
-              },
-              "required" => [
-                "steps",
-                "answer"
-              ]
-              # additionalProperties is mandatory
-              # "additionalProperties" => false
-            }
-          }
-        }
-      }
-
-      assert "Invalid schema for response_format" <> _ =
-               CommonWebhook.webhook("parse_via_gpt_vision", fields)
-    end
-  end
-
-  test "parse_via_chat_gpt, failed due to empty question_text" do
-
   end
 end
